move API namespace under CloudGarage module
